Add capture button to screenshot tab
