+++ codesign/components/augmented-reality/AugmentedRealityUI.tsx
@@ -1,6 +1,5 @@
 import { StyleSheet } from 'react-native';
 import { useState } from 'react';
-import { saveToLibraryAsync, usePermissions } from 'expo-media-library';
 
 import { ImageButton } from '@/components/ui/ImageButton';
 import { ThemedText } from '@/components/ui/ThemedText';
@@ -13,10 +12,7 @@ import { Layout } from '@/constants/styles/Layout';
 import { tamuColors } from '@/constants/Colors';
 import { Border } from '@/constants/styles/Border';
 import { useAugmentedRealityContext } from '@/components/augmented-reality/AugmentedRealityProvider';
-import { TextButton } from '@/components/ui/TextButton';
-import { Colors } from '@/constants/Colors';
-
-const SCREENSHOT_WAIT_TIME = 25; // milliseconds
+import { ScreenshotCapture } from '@/components/augmented-reality/ScreenshotCapture';
 
 type AR_UI_TAB = 'INITIAL' | 'ITEM_MENU' | 'SCREEN_CAPTURE';
 
@@ -158,77 +154,6 @@ export function AugmentedRealityUI({
   );
 }
 
-type ScreenshotCaptureProps = {
-  handleSaveSuggestion: () => void;
-  setShowEntireUI: (show: boolean) => void;
-};
-
-function ScreenshotCapture({
-  handleSaveSuggestion,
-  setShowEntireUI
-}: ScreenshotCaptureProps) {
-  const [libraryStatus, requestLibraryPermission] = usePermissions();
-
-  const {
-    setNudgeTextWithReset,
-    augmentedRealitySceneRef,
-    maybeHideNudgeText
-  } = useAugmentedRealityContext();
-
-  const handleScreenshot = async () => {
-    if (!libraryStatus?.granted) {
-      const { status } = await requestLibraryPermission();
-      if (status && status !== 'granted') {
-        setNudgeTextWithReset(
-          'Camera permission is required to take a photo. Please enable access in device settings.'
-        );
-        return;
-      }
-    }
-
-    maybeHideNudgeText();
-    setShowEntireUI(false);
-    setTimeout(() => {
-      takeScreenshot();
-    }, SCREENSHOT_WAIT_TIME);
-  };
-
-  const takeScreenshot = async () => {
-    return await augmentedRealitySceneRef?.current?.capture?.().then((uri) => {
-      saveToLibraryAsync(uri)
-        .then(() => {
-          setNudgeTextWithReset('Screenshot saved');
-          setShowEntireUI(true);
-          handleSaveSuggestion();
-        })
-        .catch(() => {
-          setNudgeTextWithReset(
-            'An issue occurred while saving the screenshot.'
-          );
-          setShowEntireUI(true);
-        })
-        .catch(() => {
-          setNudgeTextWithReset(
-            'An issue occurred while taking the screenshot.'
-          );
-          setShowEntireUI(true);
-        });
-    });
-  };
-
-  return (
-    <>
-      <ThemedView style={[styles.screenCaptureContainer]}>
-        <TextButton
-          text="Take Screenshot"
-          onPress={handleScreenshot}
-          testID="screenshot-button"
-        />
-      </ThemedView>
-    </>
-  );
-}
-
 const styles = StyleSheet.create({
   backButton: {
     position: 'absolute',
@@ -262,17 +187,5 @@ const styles = StyleSheet.create({
     textAlign: 'center',
     padding: Spacing.small,
     ...Border.roundedSmall
-  },
-  screenCaptureContainer: {
-    ...Layout.row,
-    ...Layout.justifyCenter,
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    zIndex: 1,
-    height: Spacing.xxxlarge * 3.75,
-    padding: Spacing.large,
-    backgroundColor: Colors.light.augmentedRealityTransparentBackground
   }
 });
